+++ controller/transaction.go
@@ -71,6 +71,11 @@ func DeleteTransaction(c echo.Context) error {
 	if auth == false || userList.FullName != transactionList.Users {
 		return echo.NewHTTPError(http.StatusUnauthorized, "Cannot access this account")
 	}
+
+	if transactionList.Checkout == "success" {
+		return echo.NewHTTPError(http.StatusBadRequest, "Cannot edit this transaction")
+	}
+
 	transaction, err := database.DeleteTransactions(id)
 	if err != nil {
 		return c.JSON(http.StatusInternalServerError, map[string]interface{}{
@@ -96,6 +101,11 @@ func ChangeStatus(c echo.Context) error {
 	if auth == false || userList.FullName != transaction.Users {
 		return echo.NewHTTPError(http.StatusUnauthorized, "Cannot access this account")
 	}
+
+	if transaction.Checkout == "not yet" {
+		return echo.NewHTTPError(http.StatusBadRequest, "Cannot edit this transaction")
+	}
+
 	transaction.TransactionStatus = "paid"
 	c.Bind(&transaction)
 	transactionSaved, err := database.EditTransaction(transaction)
@@ -130,6 +140,11 @@ func ChangePaymentMethod(c echo.Context) error {
 	if auth == false || userList.FullName != transaction.Users {
 		return echo.NewHTTPError(http.StatusUnauthorized, "Cannot access this account")
 	}
+
+	if transaction.Checkout == "success" {
+		return echo.NewHTTPError(http.StatusBadRequest, "Cannot edit this transaction")
+	}
+
 	transaction.PaymentMethod = payment
 	c.Bind(&transaction)
 	transactionSaved, err := database.EditTransaction(transaction)
@@ -141,5 +156,35 @@ func ChangePaymentMethod(c echo.Context) error {
 	return c.JSON(http.StatusOK, map[string]interface{}{
 		"message": "success",
 		"data":    transactionSaved,
+	})
+}
+
+func TransactionCheckout(c echo.Context) error {
+	id, err := strconv.Atoi(c.Param("transactionId"))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, map[string]interface{}{
+			"message": "invalid id",
+		})
+	}
+	auth, userList := Authorized(c)
+	transactionList, err := database.GetOneTransaction(id)
+
+	if auth == false || userList.FullName != transactionList.Users {
+		return echo.NewHTTPError(http.StatusUnauthorized, "Cannot access this account")
+	}
+	if transactionList.PaymentMethodID == 99 || transactionList.TotalPrice == 0 {
+		return echo.NewHTTPError(http.StatusBadRequest, "Please fill the transaction product and payment method")
+	}
+	transactionList.Checkout = "success"
+	c.Bind(&transactionList)
+	checkoutStatus, err := database.EditTransaction(transactionList)
+	if err != nil {
+		return c.JSON(http.StatusInternalServerError, map[string]interface{}{
+			"message": "Checkout Failure",
+		})
+	}
+	return c.JSON(http.StatusOK, map[string]interface{}{
+		"message": "success",
+		"data":    checkoutStatus,
 	})
 }
